Implement HostDispatchChains

+++ go/felix/config/config_params_test.go
@@ -120,6 +120,6 @@ var _ = DescribeTable("Config parsing",
 	Entry("PrometheusMetricsEnabled", "PrometheusMetricsEnabled", "true", true),
 	Entry("PrometheusMetricsPort", "PrometheusMetricsPort", "1234", int(1234)),
 
-	Entry("FailsafeInboundHostPorts", "FailsafeInboundHostPorts", "1,2,3,4", []int{1, 2, 3, 4}),
-	Entry("FailsafeOutboundHostPorts", "FailsafeOutboundHostPorts", "1,2,3,4", []int{1, 2, 3, 4}),
+	Entry("FailsafeInboundHostPorts", "FailsafeInboundHostPorts", "1,2,3,4", []uint16{1, 2, 3, 4}),
+	Entry("FailsafeOutboundHostPorts", "FailsafeOutboundHostPorts", "1,2,3,4", []uint16{1, 2, 3, 4}),
 )
